solve part 2 of day 13

+++ src/main/java/ch/ebynaqon/aoc/aoc24/day13/Vector.java
@@ -1,4 +1,4 @@
 package ch.ebynaqon.aoc.aoc24.day13;
 
-record Vector(int x, int y) {
+record Vector(double x, double y) {
 }
